Split out form code into files
